conditional statement lab

+++ conditional-statements-lab/excellent_result.py
@@ -1,3 +1,3 @@
-grade = int(input())
+grade = float(input())
 if grade >= 5.50:
-    print('Excellent!')
+    print("Excellent!")
